+++ src/Components/Button/Button.module.scss
@@ -1,8 +1,12 @@
-.Button__Wrapper {
+.holee-button {
   background-color: transparent;
-  width: 5rem;
-  height: 2rem;
   border-radius: 0.3rem;
   margin: 1rem;
   border: 1px solid black;
+  cursor: pointer;
+  background-color: transparent;
+}
+
+.holee-button:hover {
+  background-color: #e2e8f0;
 }
