--- src/Components/Card/Card.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import styles from './Card.module.scss';
-
-const Card = () => {
-  return <div className={styles.Card__Wrapper}>Color div</div>;
-};
-
-export default Card;
