+++ src/Components/Card/index.tsx
@@ -1,2 +1,6 @@
-import Card from './Card';
-export default Card;
+import React from 'react';
+import styles from './Card.module.scss';
+
+export const Card = () => {
+  return <div className={styles.Card__Wrapper}>Color div</div>;
+};
