+++ src/Components/Input/Input.module.scss
@@ -1,0 +1,8 @@
+.Input__Wrapper {
+  background-color: #d2d2d2;
+  width: 17rem;
+  height: 11rem;
+  border-radius: 1rem;
+  box-shadow: 3px 3px 5px rgb(0 0 0 / 25%);
+  margin: 1rem;
+}
