+++ src/Components/Input/Input.stories.tsx
@@ -1,0 +1,4 @@
+import { storiesOf } from '@storybook/react';
+import { withKnobs } from '@storybook/addon-knobs';
+
+storiesOf('Input', module).addDecorator(withKnobs);
