+++ src/Components/Input/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import styles from './Input.module.scss';
+
+export const Input = () => {
+  return <div className={styles.Input__Wrapper}>Color div</div>;
+};
