+++ src/Components/index.tsx
@@ -1,1 +1,6 @@
-export { ColorButton } from './ColorButton';
+import { ColorButton } from './ColorButton';
+import { Card } from './Card';
+import { Button } from './Button';
+import { Input } from './Input';
+
+export { ColorButton, Card, Button, Input };
