+++ src/Route/CardForm/CardForm.css
@@ -8,23 +8,26 @@
 
 .card-form__header {
   align-self: flex-start;
-  margin: 2rem 1.3rem 1.3rem 1.6rem;
+  margin-top: 1.2rem;
+  margin-left: 1.2rem;
 }
 
 .card-form__header-title {
-  padding-left: 1rem;
 }
 
 .card-form__header-button {
-  cursor: pointer;
   border: none;
-  background-color: transparent;
 }
 
 .card-form__body {
-  padding: 1rem;
+  /* padding: 1rem; */
 }
 
 .card-form__footer {
   padding: 1rem;
+}
+
+.card-form__footer-button {
+  width: 5rem;
+  height: 2rem;
 }
