+++ src/Route/CardForm/CardFormPresenter.tsx
@@ -1,13 +1,16 @@
 import React from 'react';
-import Card from '../../Components/Card';
-import Button from '../../Components/Button';
+import { Card, Button } from '../../Components';
 import './CardForm.css';
+
+const hi = () => {
+  console.log('hi');
+};
 
 const CardFormPresenter = () => {
   return (
     <div className="card-form">
       <div className="card-form__header">
-        <Button className="card-form__header-button" type="button">
+        <Button className="card-form__header-button" type="button" onClick={hi}>
           {'<'}
         </Button>
         <span className="card-form__header-title">카드 추가</span>
@@ -16,7 +19,9 @@ const CardFormPresenter = () => {
         <Card></Card>
       </div>
       <div className="card-form__footer">
-        <Button className="button-form">제출 완료</Button>
+        <Button className="card-form__footer-button" type="submit">
+          제출 완료
+        </Button>
       </div>
     </div>
   );
